feat: new default for pwd generation and online mode created
