+++ internal/cli/flags.go
@@ -10,11 +10,11 @@ import (
 )
 
 func ParseFlags(input *Input) {
-	flag.BoolVar(&input.Flags.UseOnlyNums, "numeric", false, "Password with numbers only")
-	flag.BoolVar(&input.Flags.PrintCredits, "credits", false, "Print random.org API credits to the user")
 	flag.BoolVar(&input.Flags.NeedHelp, "help", false, "Help the user to use the CLI tool")
+	flag.BoolVar(&input.Flags.UseOnlyNums, "numeric", false, "Password with numbers only")
+	flag.BoolVar(&input.Flags.Online, "online", false, "Generate random password using random.org API")
+	flag.BoolVar(&input.Flags.PrintCredits, "credits", false, "Print random.org API credits to the user")
 	flag.BoolVar(&input.Flags.UseStandard, "standard", false, "Generate password with one upper case, one number and one special character at least.")
-	flag.BoolVar(&input.Flags.Offline, "offline", false, "Generate password without using random.org API")
 	flag.IntVar(&input.NumberOfPasswords, "bulk", 1, "Size of the password")
 
 	flag.Parse()
