+++ internal/generator/mode.go
@@ -4,16 +4,16 @@ import "github.com/lucasloureiror/AegisPass/internal/cli"
 
 func ReturnGeneratorMode(data *cli.Input) PasswordGeneratorStrategy {
 
-	if data.Flags.Offline {
-		return offline{}
+	if data.Flags.NeedHelp {
+		return help{}
 	}
 
 	if data.Flags.UseStandard {
 		return standard{}
 	}
 
-	if data.Flags.NeedHelp {
-		return help{}
+	if data.Flags.Online {
+		return online{}
 	}
 
 	return random{}
